change balance to Float

+++ app/src/main/java/com/crypto/application/common/db/TransactionsEntity.kt
@@ -8,7 +8,7 @@ import java.util.Calendar
 @Entity(tableName = "transactions")
 data class TransactionsEntity(
     @PrimaryKey val time: Calendar,
-    val amount: Int,
+    val amount: Float,
     val category: String? = null,
     val transactionType: TransactionType
 )
